requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.0.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.0
azure-identity==1.6.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blobfile==2.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
duckdb==0.9.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.3
Faker==19.13.0
fastapi==0.104.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.0.352
langchain-cli==0.0.9
langchain-community==0.0.6
langchain-core==0.1.10
langchain-experimental==0.0.47
langchain-openai==0.0.2.post1
langchainhub==0.1.14
langsmith==0.0.92
llama_index==0.4.33
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
msal==1.27.0
msal-extensions==0.3.1
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.7.2
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandasai==1.5.20
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.14
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.51
stack-data==0.6.3
starlette==0.27.0
streamlit==1.20.0
streamlit-chat==0.0.2.2
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.3.0
tiktoken==0.5.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.10.0
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.23.2
validators==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
yarl==1.9.4
zipp==3.17.0
